produce/requirements.txt
kafka-python
tweepy
configparser
pymongo
pyspark